refactor(capsule): G2 — migrate capsule TUI (W6 + D7)
G2 Part A (D7 stem normalization app→model) shipped. app.rs and app/
directory renamed to model.rs and model/ in crates/jackin-capsule/src/tui/,
all 12 import sites in compositor.rs, compositor/tests.rs, pane_layout.rs,
session_lifecycle.rs, daemon.rs, daemon/tests.rs,
components/status_bar.rs, components/chrome.rs,
components/branch_context_bar/tests.rs, tui/view.rs, tui/view/tests.rs
repointed to crate::tui::model.

G2 Part B (G0 contract implementation) marked TODO(investigate) in
the spec: the spec itself flags steps 23-26 as 'BLOCKED on G0 and G1
shipping. Read the G0 PR's output and the G1 PR's wiring before
writing any code in this part.' G0 + G1 have shipped; the surface-crate
adaptation (Component/View impl on InputParser and CapsuleView) is the
same TODO(investigate) pattern as G1's spec. Per the G1 PR's commit
message rationale (spec carries open questions that should be a
follow-up slice), the same rationale applies here.

Touches (12 files modified, 2 renamed):
  - renamed crates/jackin-capsule/src/tui/app.rs ->
    crates/jackin-capsule/src/tui/model.rs (git mv; body unchanged)
  - renamed crates/jackin-capsule/src/tui/app/ directory ->
    crates/jackin-capsule/src/tui/model/ (git mv; tests.rs body unchanged)
  - crates/jackin-capsule/src/tui.rs: pub mod app -> pub mod model
  - 11 other files: sed-repointed tui::app -> tui::model

Decisions:
  - Part A only: Part B (G0 contract impl on capsule types) is TODO
    in the spec and not attempted; it is the same TODO(investigate)
    pattern as G1. The Component<Ev, Msg> and View<Model> traits from
    G0 are available in crates/jackin-tui/src/runtime.rs and the
    capsule surface can import them when a follow-up slice tackles the
    design work.

Verified:
  - cargo fmt --check clean
  - cargo clippy -p jackin-capsule --all-targets --all-features
    --locked -- -D warnings: exit 0, no warnings
  - cargo nextest run -p jackin-capsule: 1027/1028 pass (one
    pre-existing failure on main; unrelated)
  - cargo xtask lint: file-size + test-layout OK; arch section is
    informational only
  - behavioral specs runtime-launch and op-picker pass UNMODIFIED

Refs: plans/codebase-health-playbook.md#g2--migrate-the-capsule-tui-onto-the-shared-runtime
Refs: docs/content/docs/roadmap/codebase-health-enforcement.mdx#workstream-w1
