+++ crates/jackin-capsule/src/daemon.rs
@@ -71,7 +71,7 @@ use crate::session::{
 };
 use crate::socket;
 use crate::token_monitor::{TokenMonitor, TokenTotals};
-use crate::tui::app::{
+use crate::tui::model::{
     ChromeHitState, CursorVisibilityState, DragState, HoverState, HoverTarget, MuxMode,
     MuxModeState, PointerShape, PointerShapeState, VisiblePane, chrome_hover_target_for_state,
     cursor_visible_for_state, hover_target_for_state, mux_mode_for_state, pointer_shape_for_state,
