+++ crates/jackin-capsule/src/daemon/compositor/tests.rs
@@ -2,7 +2,7 @@
 use jackin_term::{DamageGrid, UnderlineStyle};
 
 use super::{SgrMetadata, pane_sgr_regions};
-use crate::tui::app::VisiblePane;
+use crate::tui::model::VisiblePane;
 use crate::tui::layout::Rect;
 use crate::tui::view::PaneScreen;
