+++ crates/jackin-capsule/src/tui.rs
@@ -4,7 +4,7 @@
 //! terminal UI lives here, per the TUI source-location convention in
 //! `tui-design-decisions.mdx`.
 
-pub mod app;
+pub mod model;
 pub mod components;
 pub mod effect;
 pub(crate) mod host_colors;
