+++ crates/jackin-capsule/src/tui/model.rs
