+++ crates/jackin-capsule/src/tui/model/tests.rs
