+++ crates/jackin-capsule/tests/status_bar.rs
@@ -1,4 +1,4 @@
-use jackin_capsule::tui::app::VisibleAgentState;
+use jackin_capsule::tui::model::VisibleAgentState;
 use jackin_capsule::tui::components::chrome::StatusBarWidget;
 use jackin_capsule::tui::components::status_bar::{PrefixMode, StatusBar, status_bar_plan};
 /// Status bar layout regressions: brand pill, tab click regions,
